# openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AnalyticsPeriod": {
        "type": "string",
        "enum": [
          "day",
          "week",
          "month"
        ],
        "default": "day"
      },
...
      "CreateExpenseRequest": {
        "type": "object",
        "required": [
          "categoryId",
          "amount"
        ],
        "properties": {
          "branchId": {
            "type": "string",
            "format": "uuid",
            "description": "Ignored on create; expense is recorded for authenticated user's branch"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of an active ExpenseCategory"
          },
          "amount": {
            "type": "number",
            "example": 3500000,
            "description": "Amount in UZS equivalent"
          },
          "currency": {
            "type": "string",
            "enum": [
              "UZS",
              "USD"
            ],
            "default": "UZS"
          },
          "amountUsd": {
            "type": "number",
            "example": 275,
            "description": "Original USD amount when currency is USD"
          },
          "usdToUzsRate": {
            "type": "number",
            "example": 12650,
            "description": "Required when currency is USD"
          },
          "description": {
            "type": "string",
            "example": "Monthly rent payment"
          },
          "expenseDate": {
            "type": "string",
            "format": "date-time",
            "description": "Defaults to now if omitted"
          }
        }
      },
...
      "InventoryRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "string",
            "description": "Decimal — current stock level",
            "example": "95.0000"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "branch": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "product": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "sku": {
                "type": "string"
              },
              "unit": {
                "type": "string"
              },
              "lowStockThreshold": {
                "type": "string",
                "nullable": true
              }
            }
          }
        }
      },
...
      "SaleType": {
        "type": "string",
        "enum": [
          "RETAIL",
          "WHOLESALE"
        ]
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "CASH_UZS",
          "CASH_USD",
          "CARD",
          "TRANSFER",
          "MIXED",
          "CREDIT"
        ]
      },
      "SaleItemRequest": {
        "type": "object",
        "required": [
          "productId",
          "quantity"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "example": 5
          }
        }
      },
      "CreateSaleRequest": {
        "type": "object",
        "required": [
          "saleType",
          "items",
          "paymentMethod"
        ],
        "properties": {
          "branchId": {
            "type": "string",
            "format": "uuid",
            "description": "Optional. Defaults to the authenticated user's branch; ADMIN cannot override it."
          },
          "customerId": {
            "type": "string",
            "format": "uuid",
            "description": "Required when paidAmountUzs < totalAmount (creates debt)"
          },
          "saleType": {
            "$ref": "#/components/schemas/SaleType"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItemRequest"
            }
          },
          "paidAmountUzs": {
            "type": "number",
            "default": 0,
            "example": 200000
          },
          "paidAmountUsd": {
            "type": "number",
            "default": 0,
            "example": 0
          },
          "usdToUzsRate": {
            "type": "number",
            "description": "Required when paidAmountUsd > 0",
            "example": 12700
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "note": {
            "type": "string"
          }
        }
      },
...
      "SaleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "saleType": {
            "$ref": "#/components/schemas/SaleType"
          },
          "totalAmountUzs": {
            "type": "string",
            "example": "250000.00"
          },
          "paidAmountUzs": {
            "type": "string",
            "example": "200000.00"
          },
          "debtAmountUzs": {
            "type": "string",
            "example": "50000.00"
          },
          "branch": {
            "type": "object"
          },
          "customer": {
            "type": "object",
            "nullable": true
          },
          "soldBy": {
            "type": "object"
          },
          "items": {
            "type": "array"
          },
          "payments": {
            "type": "array"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "openapi": "3.0.0",
  "info": {
    "title": "Store Management API",
    "version": "1.0.0",
    "description": "Multi-branch retail management system API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/analytics/dashboard": {
      "get": {
        "summary": "Top-level KPIs for a date range",
        "description": "Returns aggregated sales, expenses, net profit, inventory value,\nlow-stock count, customer debt, and pending transfers.\nADMIN sees their branch only; SUPER_ADMIN sees all or a specific branch.\n",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "SUPER_ADMIN only"
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Defaults to the first day of the current month"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Defaults to now"
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard KPIs"
          }
        }
      }
    },
    "/analytics/sales": {
      "get": {
        "summary": "Sales breakdown — by period, type, payment method, and top products",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "period",
            "schema": {
              "$ref": "#/components/schemas/AnalyticsPeriod"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            },
            "description": "Number of top products to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Sales report"
          }
        }
      }
    },
    "/analytics/inventory": {
      "get": {
        "summary": "Inventory report — stock value per branch, low-stock items, movement summary",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Used to scope the stock-movement summary"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory report"
          }
        }
      }
    },
    "/analytics/expenses": {
      "get": {
        "summary": "Expense breakdown — totals by category and over time",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "period",
            "schema": {
              "$ref": "#/components/schemas/AnalyticsPeriod"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense report"
          }
        }
      }
    },
    "/analytics/customers/debt": {
      "get": {
        "summary": "Customer debt summary — total outstanding, overdue breakdown, and top debtors",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer debt report"
          }
        }
      }
    },
...
    "/expenses": {
      "post": {
        "summary": "Record a branch expense",
        "tags": [
          "Expenses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense recorded"
          },
          "404": {
            "description": "Expense category not found"
          },
          "409": {
            "description": "Category is inactive"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "List expenses (branch-scoped for ADMIN)",
        "tags": [
          "Expenses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense list"
          }
        }
      }
    },
...
    "/inventory": {
      "get": {
        "summary": "Current stock levels (branch-scoped for ADMIN)",
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "SUPER_ADMIN only — omit to see all branches"
          },
          {
            "in": "query",
            "name": "productId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "lowStock",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter products at or below their lowStockThreshold"
          }
        ],
        "responses": {
          "200": {
            "description": "List of inventory records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/sales": {
      "post": {
        "summary": "Create a sale (deducts inventory + updates customer debt atomically)",
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sale recorded with inventory deducted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Debt requires a customer"
          },
          "404": {
            "description": "Product, branch or customer not found"
          },
          "409": {
            "description": "Insufficient stock or inactive product"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "List sales with filters (branch-scoped for ADMIN)",
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "customerId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "saleType",
            "schema": {
              "$ref": "#/components/schemas/SaleType"
            }
          },
          {
            "in": "query",
            "name": "hasDebt",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter sales with outstanding debt"
          },
          {
            "in": "query",
            "name": "overdue",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter sales where debtDueDate has passed and debt remains"
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sale list (without items/payments for performance)"
          }
        }
      }
    },
...
  },
...
}